% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{clear_venv_path}
\alias{clear_venv_path}
\title{Remove the saved virtualenv path}
\usage{
clear_venv_path()
}
\value{
No return value, called for side effects.
}
\description{
Remove the saved virtualenv path
}
\keyword{internal}
