% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{config_dir}
\alias{config_dir}
\title{Path to the package config directory}
\usage{
config_dir()
}
\value{
Character path to the config directory.
}
\description{
Path to the package config directory
}
\keyword{internal}
