% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{find_free_port}
\alias{find_free_port}
\title{Find a free TCP port}
\usage{
find_free_port()
}
\value{
Integer port number.
}
\description{
Samples random ports in the dynamic range and uses \code{\link[=serverSocket]{serverSocket()}} to
verify availability.
}
\keyword{internal}
