% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{load_venv_path}
\alias{load_venv_path}
\title{Load the saved virtualenv path (or NULL)}
\usage{
load_venv_path()
}
\value{
Character path or \code{NULL}.
}
\description{
Load the saved virtualenv path (or NULL)
}
\keyword{internal}
