% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{save_venv_path}
\alias{save_venv_path}
\title{Save the virtualenv path to persistent config}
\usage{
save_venv_path(path)
}
\arguments{
\item{path}{Character path to save.}
}
\value{
No return value, called for side effects.
}
\description{
Save the virtualenv path to persistent config
}
\keyword{internal}
