% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{stop_server}
\alias{stop_server}
\title{Stop the citest API server}
\usage{
stop_server()
}
\value{
No return value, called for side effects.
}
\description{
Kills the background Python process and clears the internal state.
}
\examples{
\dontshow{if (citestR:::has_server()) withAutoprint(\{ # examplesIf}
stop_server()
\dontshow{\}) # examplesIf}
}
