% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{to_nested_list}
\alias{to_nested_list}
\title{Convert an R matrix / data.frame to a nested list suitable for JSON}
\usage{
to_nested_list(x)
}
\description{
Convert an R matrix / data.frame to a nested list suitable for JSON
}
\keyword{internal}
